refactor: Cleanup pages
